fix(cpu/rmsnorm): accumulate sum of squares in f64 for numerical precision
Use f64 for the sum-of-squares accumulator and inverse RMS computation
across all CPU SIMD backends (scalar, AVX2, AVX-512, NEON). This avoids
f32 accumulation overflow and precision loss for large hidden dimensions,
matching the behavior of llama.cpp's ggml_float accumulator.

For AVX2, the 8-lane f32 vector is split into two 4-lane f64 vectors
before squaring and accumulating, keeping the hot SIMD loop in f64
throughout. AVX-512 and NEON promote the f32 horizontal-sum result to
f64 before the scalar tail and final division.

+++ src/runtime/cpu/kernels/simd/norm/aarch64/neon/rms_norm.rs
@@ -35,16 +35,16 @@ pub unsafe fn rms_norm_f32(
             let v = vld1q_f32(base.add(i * F32_LANES));
             ss_acc = vfmaq_f32(ss_acc, v, v);
         }
-        let mut sum_sq = hsum_f32(ss_acc);
+        let mut sum_sq = hsum_f32(ss_acc) as f64;
 
         // Scalar tail for sum of squares
         for i in 0..remainder {
-            let v = *base.add(chunks * F32_LANES + i);
+            let v = *base.add(chunks * F32_LANES + i) as f64;
             sum_sq += v * v;
         }
 
-        // Compute inverse RMS: 1 / sqrt(mean_sq + eps)
-        let inv_rms = 1.0 / (sum_sq / hidden_size as f32 + eps).sqrt();
+        // Compute inverse RMS in f64 for precision (matches llama.cpp)
+        let inv_rms = (1.0f64 / (sum_sq / hidden_size as f64 + eps as f64).sqrt()) as f32;
         let v_inv_rms = vdupq_n_f32(inv_rms);
 
         // Phase 2: Apply normalization and weight
